requirements.txt
-i https://pypi.org/simple
aiofiles==23.2.1
annotated-types==0.5.0 ; python_version >= '3.7'
anyio==3.7.1 ; python_version >= '3.7'
certifi==2023.7.22 ; python_version >= '3.6'
charset-normalizer==3.2.0
click==8.1.7 ; python_version >= '3.7'
cmake==3.27.4.1
contourpy==1.1.0 ; python_version >= '3.8'
cycler==0.11.0 ; python_version >= '3.6'
deprecation==2.1.0
fastapi==0.103.1
filelock==3.12.3 ; python_version >= '3.8'
fonttools==4.42.1 ; python_version >= '3.8'
fsspec==2023.9.0 ; python_version >= '3.8'
gensim==4.3.2
gotrue==1.1.0 ; python_version >= '3.8' and python_version < '4.0'
h11==0.14.0 ; python_version >= '3.7'
httpcore==0.17.3 ; python_version >= '3.7'
httptools==0.6.0
httpx==0.24.1 ; python_version >= '3.7'
huggingface-hub==0.16.4 ; python_full_version >= '3.7.0'
idna==3.4 ; python_version >= '3.5'
jinja2==3.1.2
joblib==1.3.2 ; python_version >= '3.7'
kiwisolver==1.4.5 ; python_version >= '3.7'
lit==16.0.6
markupsafe==2.1.3 ; python_version >= '3.7'
matplotlib==3.7.2 ; python_version >= '3.8'
mpmath==1.3.0
networkx==3.1 ; python_version >= '3.8'
numpy==1.25.2
nvidia-cublas-cu11==11.10.3.66 ; platform_system == 'Linux' and platform_machine == 'x86_64'
nvidia-cuda-cupti-cu11==11.7.101 ; platform_system == 'Linux' and platform_machine == 'x86_64'
nvidia-cuda-nvrtc-cu11==11.7.99 ; platform_system == 'Linux' and platform_machine == 'x86_64'
nvidia-cuda-runtime-cu11==11.7.99 ; platform_system == 'Linux' and platform_machine == 'x86_64'
nvidia-cudnn-cu11==8.5.0.96 ; platform_system == 'Linux' and platform_machine == 'x86_64'
nvidia-cufft-cu11==10.9.0.58 ; platform_system == 'Linux' and platform_machine == 'x86_64'
nvidia-curand-cu11==10.2.10.91 ; platform_system == 'Linux' and platform_machine == 'x86_64'
nvidia-cusolver-cu11==11.4.0.1 ; platform_system == 'Linux' and platform_machine == 'x86_64'
nvidia-cusparse-cu11==11.7.4.91 ; platform_system == 'Linux' and platform_machine == 'x86_64'
nvidia-nccl-cu11==2.14.3 ; platform_system == 'Linux' and platform_machine == 'x86_64'
nvidia-nvtx-cu11==11.7.91 ; platform_system == 'Linux' and platform_machine == 'x86_64'
orjson==3.9.7
packaging==23.1 ; python_version >= '3.7'
pandas==2.1.0
pillow==10.0.0
postgrest==0.10.8 ; python_version >= '3.8' and python_version < '4.0'
pyclustering==0.10.1.2
pydantic==2.3.0
pydantic-core==2.6.3 ; python_version >= '3.7'
pymystem3==0.2.0
pyparsing==3.0.9 ; python_full_version >= '3.6.8'
python-dateutil==2.8.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
pyyaml==6.0.1 ; python_version >= '3.6'
realtime==1.0.0 ; python_version >= '3.8' and python_version < '4.0'
regex==2023.8.8 ; python_version >= '3.6'
requests==2.31.0 ; python_version >= '3.7'
safetensors==0.3.3
scikit-learn==1.3.0
scipy==1.11.2 ; python_version < '3.13' and python_version >= '3.9'
setuptools==68.2.0 ; python_version >= '3.8'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
smart-open==6.4.0 ; python_version >= '3.6' and python_version < '4.0'
sniffio==1.3.0 ; python_version >= '3.7'
starlette==0.27.0 ; python_version >= '3.7'
storage3==0.5.4 ; python_version >= '3.8' and python_version < '4.0'
strenum==0.4.15
supabase==1.0.4
supafunc==0.2.3 ; python_version >= '3.8' and python_version < '4.0'
sympy==1.12 ; python_version >= '3.8'
threadpoolctl==3.2.0 ; python_version >= '3.8'
tokenizers==0.13.3
torch==2.0.1
tqdm==4.66.1 ; python_version >= '3.7'
transformers==4.33.1
triton==2.0.0 ; platform_system == 'Linux' and platform_machine == 'x86_64'
typing-extensions==4.7.1 ; python_version >= '3.7'
tzdata==2023.3 ; python_version >= '2'
urllib3==2.0.4 ; python_version >= '3.7'
uvicorn[standard]==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
websockets==10.4
wheel==0.41.2 ; python_version >= '3.7'
wordcloud==1.9.2
